requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
Flask-Login==0.6.3
pymongo==4.3.3
python-dotenv==1.0.0
marshmallow==3.19.0
psycopg2==2.9.10
config==0.5.1
# testing
pytest==7.4.0
pytest-flask==1.2.0
Flask-JWT-Extended==4.7.1
